Auto join and a lot of other stuff
- Updated Readme
- Updated version to 1.0.2
- Updated version in manifest, as well as added new offline logo.
- Added functionality to update browser icon when start/stop has been
dispatched.
- Added new automation script that runs the auto join scripts, as well
as future auto-scripts that may be added.
- Added auto join script. Now when the hangout is launched after the
start time is matched, it will inject a content script into the tab’s
dom when completed, and then dispatch a click event on the join button.
Right now it uses jQuery to find the join button, but this will change
in the future.
- Added new ‘events’ class that handles the clicking of the buttons, as
I believe we will need this in the future as well when we implement the
stay alive auto-script.
- Where the lib file lives has now changed to `js/lib`
- The backgroundWorker now changes the browser icon on initial data
load if started is true.
- Start Listener has changed to include new automation.js file.
- A lot of UI changes. Background, buttons, etc.
- A new Auto Join Toggle component was added to the panel ui, you can
now enable/disable auto-join with one click.
- Some gulp updates/changes to accommodate the automation.js changes.